refactor: move admin settings controller, views, routes, and i18n to collavre engine
Move admin settings from host app to collavre core engine:
- Controller: Admin::SettingsController -> Collavre::Admin::SettingsController
- Views: app/views/admin/settings/ -> engine views
- Routes: /admin, /admin/uiux -> engine routes
- i18n: admin.en.yml, admin.ko.yml -> engine locales
- Tests: moved to engine test directory
- Updated all path helpers to use collavre.* prefix
- Host app can override via standard Rails engine override pattern
