+++ config/routes.rb
@@ -33,12 +33,7 @@ Rails.application.routes.draw do
   root "collavre/creatives#index"
 
   # Admin settings (host app specific)
-  get "/admin", to: "admin/settings#index"
-  get "/admin/uiux", to: "admin/settings#uiux", as: :admin_uiux
-  patch "/admin/uiux", to: "admin/settings#update_uiux"
-  namespace :admin do
-    resource :settings, only: [ :update ]
-  end
+  # Admin settings have been moved to the collavre engine.
   # MCP OAuth Discovery
   get "/.well-known/oauth-protected-resource", to: "mcp/discovery#oauth_protected_resource"
   get "/.well-known/oauth-authorization-server", to: "mcp/discovery#oauth_authorization_server"
