+++ engines/collavre/app/views/collavre/admin/settings/index.html.erb
@@ -7,5 +7,5 @@
 <%= render "admin/shared/tabs" %>
 
 <div class="tab-panels">
-  <%= render "admin/settings/uiux_tab" %>
+  <%= render "collavre/admin/settings/system_tab" %>
 </div>
