+++ engines/collavre/app/views/collavre/admin/settings/uiux.html.erb
@@ -7,5 +7,5 @@
 <%= render "admin/shared/tabs" %>
 
 <div class="tab-panels">
-  <%= render "admin/settings/system_tab" %>
+  <%= render "collavre/admin/settings/uiux_tab" %>
 </div>
